refactor(v0): extract session state write service
